Created rating services and methods

+++ db/migrate/20150319180204_create_outfits.rb
@@ -3,9 +3,8 @@ class CreateOutfits < ActiveRecord::Migration
     create_table :outfits do |t|
       t.string :name
       t.string :description
-      t.integer :rating
+      t.float :rating
       t.integer :user_id
-
       t.timestamps null: false
     end
   end
